requirements.txt
altgraph==0.17.4
asgiref==3.8.1
auto-py-to-exe==2.44.1
bottle==0.13.1
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1
django-cors-headers==4.5.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Eel==0.17.0
et-xmlfile==1.1.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.0
idna==3.10
numpy==2.1.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.0.3
